requirements.txt
torch
torchvision
pyyaml
wandb
matplotlib